Make separate test_standalone
